Move diff chip
