+++ packages/core-components/src/components/ContentElement/DiffChip.vue
